Auto-select first quiz if none is chosen
Added logic to automatically select the first quiz when no quiz is currently selected. This enhancement checks if `selectedQuiz` is null and if there are available quizzes, improving user experience by ensuring that users don't have to manually select a quiz. Now, the only thing left to do is to make sure the quizzes are actually worth taking!