Add Url property to UserQuiz and related updates
This commit introduces a new `Url` property to the `UserQuiz` entity, allowing quizzes to have an associated URL (max length 300 characters). It also updates the `Quiz` entity to include the same property.

Additionally, various repository methods have been modified to handle the new `Url` property, and the `PerplexityService` and `RollYourOwn.razor` files have been updated to set this property when creating quizzes from URLs.

Finally, a new migration has been created to add the `Url` column to the `UserQuizzes` table in the database, ensuring our schema is as up-to-date as your favorite meme!